+++ src/lib/mem.h
@@ -5,7 +5,7 @@
 
 #include <include/types.h>
 
-void memcpy(u8* source, u8* dest, int no_bytes);
+void memcpy(u64* source, u64* dest, int no_bytes);
 void memset(u64* dest, u8 val, u32 len);
 int  memcmp(const void* ptr1, const void* ptr2, u16 num);
 void *memmove(void *dest, const void *src, size_t n);
